added sparse layers

+++ rosenbrock/train_network.py
@@ -7,9 +7,11 @@ from dotenv import load_dotenv
 
 from rosenbrock.data_loaders.coordinate_dataset_loader import CoordinateDataset
 from rosenbrock.networks.dense_network import DenseNetwork
+from rosenbrock.networks.sparse_traingular_network import SpareTraingularNetwork
 
 load_dotenv()
 train_dense_network = True
+
 
 if os.getenv("TRAIN_DENSE_NETWORK") == "0":
     train_dense_network = False
@@ -23,32 +25,42 @@ device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
 # Hyper-parameters
 layer_size = 2
-num_epochs = 10
-batch_size = 100
-learning_rate = 0.0001
+num_epochs = 5
+batch_size = 1
+learning_rate = 0.001
 
 # training and test datasets
 train_dataset = CoordinateDataset("rosenbrock/datasets/rosenbrock_training_data.csv")
 test_dataset = CoordinateDataset("rosenbrock/datasets/rosenbrock_testing_data.csv")
 
-# normalisation
-train_loader_norm = torch.utils.data.DataLoader(
-    dataset=train_dataset, batch_size=len(train_dataset), shuffle=False
+# Normalization
+params_file_path = (
+    "rosenbrock/saved_params/rosenbrock_sampled_data_normalisation_params.pkl"
 )
 
-min_value = float("inf")
-max_value = float("-inf")
-for input_values, output_values in enumerate(train_loader_norm):
-    concat_tensor = torch.cat(output_values, dim=0)
-    min_value = torch.min(concat_tensor)
-    max_value = torch.max(concat_tensor)
+if os.path.exists(params_file_path):
+    with open(params_file_path, "rb") as f:
+        min_max_scaling_params = pickle.load(f)
+else:
+    train_loader_norm = torch.utils.data.DataLoader(
+        dataset=train_dataset, batch_size=len(train_dataset), shuffle=False
+    )
 
-min_max_scaling_params = {"min": min_value, "max": max_value}
-with open(
-    f'rosenbrock/saved_params/rosenbrock_sampled_data_{"dense" if train_dense_network else "sparse"}.pkl',
-    "wb",
-) as f:
-    pickle.dump(min_max_scaling_params, f)
+    min_value = float("inf")
+    max_value = float("-inf")
+
+    for _, (input_values, output_values) in enumerate(train_loader_norm):
+        current_min = torch.min(output_values)
+        current_max = torch.max(output_values)
+
+        min_value = min(min_value, current_min)
+        max_value = max(max_value, current_max)
+
+    min_max_scaling_params = {"min": min_value, "max": max_value}
+    print(min_max_scaling_params)
+
+    with open(params_file_path, "wb") as f:
+        pickle.dump(min_max_scaling_params, f)
 
 
 # Data loaders
@@ -56,29 +68,35 @@ train_loader = torch.utils.data.DataLoader(
     dataset=train_dataset, batch_size=batch_size, shuffle=True
 )
 
+
 test_loader = torch.utils.data.DataLoader(
     dataset=test_dataset, batch_size=batch_size, shuffle=False
 )
 
 
 # model
-if True:
+if train_dense_network:
     model = DenseNetwork(layer_size).to(device)
-# else:
-#     model = sparse_network(layer_size).to(device)
-# model.float()
+else:
+    model = SpareTraingularNetwork(layer_size).to(device)
+
+model.float()
 
 # Loss and optimizer
 criterion = nn.MSELoss()
 optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
+# scheduler = torch.optim.lr_scheduler.StepLR(optimizer, step_size=2, gamma=0.1)
 
 # Train the model
 n_total_steps = len(train_loader)
 for epoch in range(num_epochs):
+    model.train()
     for i, (input_values, output_values) in enumerate(train_loader):
         # resize
         input_values = input_values.to(device)
-        output_values = (output_values.to(device) - min_value) / (max_value - min_value)
+        output_values = (output_values.to(device) - min_max_scaling_params["min"]) / (
+            min_max_scaling_params["max"] - min_max_scaling_params["min"]
+        )
         # Forward pass
         outputs = model(input_values.float())
         loss = criterion(outputs, output_values)
@@ -93,12 +111,18 @@ for epoch in range(num_epochs):
                 f"Epoch [{epoch+1}/{num_epochs}], Step [{i+1}/{n_total_steps}], Loss: {loss.item():.8f}"
             )
 
+    # scheduler.step()
+
+
 # Test the model
 with torch.no_grad():
     total_r2 = []
     for input_values, output_values in test_loader:
         input_values = input_values.to(device)
-        outputs = (model(input_values) * (max_value - min_value)) + min_value
+        outputs = (
+            model(input_values)
+            * (min_max_scaling_params["max"] - min_max_scaling_params["min"])
+        ) + min_max_scaling_params["min"]
 
         # calculate the mean and the variance
         mean_input_values = torch.mean(input_values, dim=0)
